chore: refactor logo components assets
- the images don't need to be in a separeted folder, as the components
already describes what they are being used for
- each logo component should have its own styles file
