+++ src/components/ModelLogo/style.css
@@ -1,0 +1,4 @@
+.modelLogo {
+  backface-visibility: hidden;
+  position: absolute;
+}
